# src/applications/server/main/io/infranexum/server/dcim/DcimPhysicalController.java
package io.infranexum.server.dcim;

import io.infranexum.server.http.AuthenticatedActorContext;
import static io.infranexum.server.dcim.DcimPhysicalApiModels.*;
import io.infranexum.core.contracts.DomainIdentifier;
import io.infranexum.core.contracts.UuidV7Generator;
import io.infranexum.dcim.physical.application.*;
import io.infranexum.dcim.physical.domain.*;
import io.infranexum.identity.access.domain.AuthorizationScope;
import io.infranexum.identity.access.domain.PermissionCodes;
import io.infranexum.server.http.ApiPagination;
import io.infranexum.server.identityaccess.ScopedAuthorizationGuard;
import io.infranexum.server.observability.CorrelationContext;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import jakarta.validation.Valid;
import java.util.List;
import java.util.Objects;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

/** HTTP boundary for PGM-07-E05 models, racks, equipment, ports and cables. */
@RestController
@ConditionalOnProperty(name="infranexum.dcim.physical-api-enabled",havingValue="true")
@RequestMapping("/api/v1/dcim")
public final class DcimPhysicalController {
    private final DcimPhysicalApplicationService service; private final ScopedAuthorizationGuard authorization; private final UuidV7Generator ids;
    public DcimPhysicalController(DcimPhysicalApplicationService service,ScopedAuthorizationGuard authorization,@Qualifier("correlationIdentifiers") UuidV7Generator ids){this.service=Objects.requireNonNull(service,"service");this.authorization=Objects.requireNonNull(authorization,"authorization");this.ids=Objects.requireNonNull(ids,"ids");}
    @GetMapping("/equipment-models") ResponseEntity<List<ModelResponse>> models(@RequestParam(name="organization_id") String org,@RequestParam(defaultValue="0") int offset,@RequestParam(defaultValue="100") int limit,HttpServletRequest req,HttpServletResponse res){DomainIdentifier o=id(org);require(req,res,PermissionCodes.DCIM_MODEL_READ,o,"dcim-model","collection");var page=service.models(o,offset,limit);return ApiPagination.offset(page.items().stream().map(ModelResponse::from).toList(),page.nextOffset(),limit);}
    @GetMapping("/equipment-taxonomy") ResponseEntity<EquipmentTaxonomyResponse> taxonomy(@RequestParam(name="organization_id") String org,HttpServletRequest req,HttpServletResponse res){DomainIdentifier o=id(org);require(req,res,PermissionCodes.DCIM_MODEL_READ,o,"dcim-model","taxonomy");return ResponseEntity.ok(EquipmentTaxonomyResponse.create());}
    @PostMapping("/equipment-models") ResponseEntity<ModelResponse> createModel(@RequestHeader("Idempotency-Key") String idem,@Valid @RequestBody CreateModelRequest body,HttpServletRequest req,HttpServletResponse res){DomainIdentifier org=id(body.organizationId());require(req,res,PermissionCodes.DCIM_MODEL_CREATE,org,"dcim-model","collection");var templates=(body.portTemplates()==null?List.<PortTemplateRequest>of():body.portTemplates()).stream().map(x->new PortTemplate(x.namePrefix(),x.count(),PortKind.valueOf(x.kind().strip().toUpperCase()),x.media(),x.connector())).toList();var category=EquipmentCategory.parse(body.category());var equipmentType=EquipmentType.parse(body.equipmentType());var result=service.createModel(new CreateEquipmentModelCommand(org,id(body.manufacturerPartnerId()),body.code(),body.displayName(),category,equipmentType,body.manufacturerReference(),body.formFactor(),body.rackUnits(),body.widthMm(),body.depthMm(),body.weightKg(),templates,body.description()),idem,context(req,body.reason()));return response(ModelResponse.from(result),result.version(),HttpStatus.CREATED);}
    @PostMapping("/equipment-models/{id}/status") ResponseEntity<ModelResponse> modelStatus(@PathVariable String id,@RequestHeader(HttpHeaders.IF_MATCH) String etag,@RequestHeader("Idempotency-Key") String idem,@Valid @RequestBody StatusRequest body,HttpServletRequest req,HttpServletResponse res){var current=service.model(id(id));require(req,res,PermissionCodes.DCIM_MODEL_UPDATE,current.organizationId(),"dcim-model",current.id().toString());var changed=service.changeModelStatus(current.id(),version(etag),PhysicalStatus.parse(body.targetStatus()),idem,context(req,body.reason()));return response(ModelResponse.from(changed),changed.version(),HttpStatus.OK);}
    @GetMapping("/racks") ResponseEntity<List<RackResponse>> racks(@RequestParam(name="organization_id") String org,@RequestParam(name="room_id",required=false) String room,@RequestParam(required=false) String cursor,@RequestParam(defaultValue="100") int limit,HttpServletRequest req,HttpServletResponse res){DomainIdentifier o=id(org);require(req,res,PermissionCodes.DCIM_RACK_READ,o,"dcim-rack","collection");var page=service.racks(o,nullableId(room),nullableId(cursor),limit);return ApiPagination.cursor(page.items().stream().map(RackResponse::from).toList(),text(page.nextCursor()),limit);}
    @PostMapping("/racks") ResponseEntity<RackResponse> createRack(@RequestHeader("Idempotency-Key") String idem,@Valid @RequestBody CreateRackRequest body,HttpServletRequest req,HttpServletResponse res){DomainIdentifier org=id(body.organizationId());require(req,res,PermissionCodes.DCIM_RACK_CREATE,org,"dcim-rack","collection");var r=service.createRack(new CreateRackCommand(org,id(body.subdivisionId()),id(body.roomId()),body.code(),body.displayName(),body.heightU(),body.widthMm(),body.depthMm()),idem,context(req,body.reason()));return response(RackResponse.from(r),r.version(),HttpStatus.CREATED);}
    @PostMapping("/racks/{id}/status") ResponseEntity<RackResponse> rackStatus(@PathVariable String id,@RequestHeader(HttpHeaders.IF_MATCH) String etag,@RequestHeader("Idempotency-Key") String idem,@Valid @RequestBody StatusRequest body,HttpServletRequest req,HttpServletResponse res){var current=service.rack(id(id));String permission=PhysicalStatus.parse(body.targetStatus())==PhysicalStatus.DECOMMISSIONED?PermissionCodes.DCIM_RACK_DECOMMISSION:PermissionCodes.DCIM_RACK_UPDATE;require(req,res,permission,current.organizationId(),"dcim-rack",current.id().toString());var changed=service.changeRackStatus(current.id(),version(etag),PhysicalStatus.parse(body.targetStatus()),idem,context(req,body.reason()));return response(RackResponse.from(changed),changed.version(),HttpStatus.OK);}
    @GetMapping("/equipment") ResponseEntity<List<EquipmentResponse>> equipment(@RequestParam(name="organization_id") String org,@RequestParam(name="rack_id",required=false) String rack,@RequestParam(required=false) String cursor,@RequestParam(defaultValue="100") int limit,HttpServletRequest req,HttpServletResponse res){DomainIdentifier o=id(org);require(req,res,PermissionCodes.DCIM_EQUIPMENT_READ,o,"dcim-equipment","collection");var page=service.equipment(o,nullableId(rack),nullableId(cursor),limit);return ApiPagination.cursor(page.items().stream().map(EquipmentResponse::from).toList(),text(page.nextCursor()),limit);}
    @PostMapping("/equipment") ResponseEntity<EquipmentResponse> install(@RequestHeader("Idempotency-Key") String idem,@Valid @RequestBody InstallEquipmentRequest body,HttpServletRequest req,HttpServletResponse res){DomainIdentifier org=id(body.organizationId());require(req,res,PermissionCodes.DCIM_EQUIPMENT_CREATE,org,"dcim-equipment","collection");var e=service.install(new InstallEquipmentCommand(org,id(body.subdivisionId()),id(body.rackId()),id(body.modelId()),id(body.rsotObjectId()),nullableId(body.itamAssetId()),body.serialNumber(),body.assetTag(),body.startU(),body.face()),idem,context(req,body.reason()));return response(EquipmentResponse.from(e),e.version(),HttpStatus.CREATED);}
    @PostMapping("/equipment/{id}/move") ResponseEntity<EquipmentResponse> move(@PathVariable String id,@RequestHeader(HttpHeaders.IF_MATCH) String etag,@RequestHeader("Idempotency-Key") String idem,@Valid @RequestBody MoveEquipmentRequest body,HttpServletRequest req,HttpServletResponse res){var current=service.equipment(id(id));require(req,res,PermissionCodes.DCIM_EQUIPMENT_MOVE,current.organizationId(),"dcim-equipment",current.id().toString());var e=service.moveEquipment(current.id(),version(etag),id(body.rackId()),body.startU(),body.face(),idem,context(req,body.reason()));return response(EquipmentResponse.from(e),e.version(),HttpStatus.OK);}
    @PostMapping("/equipment/{id}/status") ResponseEntity<EquipmentResponse> equipmentStatus(@PathVariable String id,@RequestHeader(HttpHeaders.IF_MATCH) String etag,@RequestHeader("Idempotency-Key") String idem,@Valid @RequestBody StatusRequest body,HttpServletRequest req,HttpServletResponse res){var current=service.equipment(id(id));PhysicalStatus target=PhysicalStatus.parse(body.targetStatus());String permission=(target==PhysicalStatus.DECOMMISSIONED||target==PhysicalStatus.ARCHIVED)?PermissionCodes.DCIM_EQUIPMENT_DECOMMISSION:PermissionCodes.DCIM_EQUIPMENT_UPDATE;require(req,res,permission,current.organizationId(),"dcim-equipment",current.id().toString());var e=service.changeEquipmentStatus(current.id(),version(etag),target,idem,context(req,body.reason()));return response(EquipmentResponse.from(e),e.version(),HttpStatus.OK);}
    @GetMapping("/equipment/{id}/ports") ResponseEntity<List<PortResponse>> ports(@PathVariable String id,@RequestParam(defaultValue="0") int offset,@RequestParam(defaultValue="100") int limit,HttpServletRequest req,HttpServletResponse res){var current=service.equipment(id(id));require(req,res,PermissionCodes.DCIM_PORT_READ,current.organizationId(),"dcim-port",current.id().toString());var page=service.ports(current.id(),offset,limit);return ApiPagination.offset(page.items().stream().map(PortResponse::from).toList(),page.nextOffset(),limit);}
    @GetMapping("/cables") ResponseEntity<List<CableResponse>> cables(@RequestParam(name="organization_id") String org,@RequestParam(required=false) String cursor,@RequestParam(defaultValue="100") int limit,HttpServletRequest req,HttpServletResponse res){DomainIdentifier o=id(org);require(req,res,PermissionCodes.DCIM_CABLE_READ,o,"dcim-cable","collection");var page=service.cables(o,nullableId(cursor),limit);return ApiPagination.cursor(page.items().stream().map(CableResponse::from).toList(),text(page.nextCursor()),limit);}
    @PostMapping("/cables") ResponseEntity<CableResponse> connect(@RequestHeader("Idempotency-Key") String idem,@Valid @RequestBody ConnectPortsRequest body,HttpServletRequest req,HttpServletResponse res){DomainIdentifier org=id(body.organizationId());require(req,res,PermissionCodes.DCIM_CABLE_CREATE,org,"dcim-cable","collection");var c=service.connect(new ConnectPortsCommand(org,id(body.subdivisionId()),id(body.portAId()),id(body.portBId()),body.label(),CableType.parse(body.cableType()),body.lengthMeters()==null?java.math.BigDecimal.ONE:body.lengthMeters(),nullableId(body.manufacturerPartnerId()),body.manufacturerReference()),idem,context(req,body.reason()));return response(CableResponse.from(c),c.version(),HttpStatus.CREATED);}
    @PostMapping("/cables/{id}/disconnect") ResponseEntity<CableResponse> disconnect(@PathVariable String id,@RequestHeader(HttpHeaders.IF_MATCH) String etag,@RequestHeader("Idempotency-Key") String idem,@Valid @RequestBody StatusRequest body,HttpServletRequest req,HttpServletResponse res){var current=service.cable(id(id));require(req,res,PermissionCodes.DCIM_CABLE_DISCONNECT,current.organizationId(),"dcim-cable",current.id().toString());var c=service.disconnect(current.id(),version(etag),idem,context(req,body.reason()));return response(CableResponse.from(c),c.version(),HttpStatus.OK);}
    private void require(HttpServletRequest req,HttpServletResponse res,String permission,DomainIdentifier org,String type,String target){authorization.require(req,res,permission,AuthorizationScope.organization(org),type,target);} private PhysicalCommandContext context(HttpServletRequest req,String reason){Object actor=req.getAttribute(AuthenticatedActorContext.ACCOUNT_ATTRIBUTE);if(!(actor instanceof DomainIdentifier actorId))throw new IllegalStateException("authenticated actor missing after RBAC boundary");return new PhysicalCommandContext(actorId,CorrelationContext.identifier(req).orElseGet(ids::next),reason);} private static DomainIdentifier id(String value){return DomainIdentifier.parse(value);} private static DomainIdentifier nullableId(String value){return value==null||value.isBlank()?null:id(value);} private static String text(DomainIdentifier value){return value==null?null:value.toString();} private static long version(String raw){if(raw==null||raw.isBlank())throw new IllegalArgumentException("If-Match is required");String v=raw.strip().replace("W/","").replace("\"","");if(v.startsWith("ver-"))v=v.substring(4);long n=Long.parseLong(v);if(n<1)throw new IllegalArgumentException("If-Match version must be positive");return n;} private static <T>ResponseEntity<T> response(T body,long version,HttpStatus status){return ResponseEntity.status(status).eTag("\"ver-"+version+"\"").body(body);}
}