# src/applications/server/resources/openapi/dcim-physical.yaml
openapi: 3.1.0
info:
  title: InfraNexum DCIM Physical API
  version: 2.0.0-alpha.0.123
  description: PGM-07-E05 rack occupancy, equipment footprints, physical ports and point-to-point cabling.
...
paths:
  /api/v1/dcim/equipment-models:
    get:
      tags:
      - DCIM / Physical / Models
      operationId: listDcimEquipmentModels
      x-infranexum-capability: dcim.physical
      summary: List governed equipment models
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - $ref: '#/components/parameters/Limit'
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          maximum: 1000000
      responses:
        '200':
          description: Equipment models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipmentModel'
          headers:
            X-Page-Limit:
              description: Effective bounded page size.
              schema:
                type: integer
                minimum: 1
            X-Next-Offset:
              description: Offset for the next page; absent when the current page is terminal.
              schema:
                type: integer
                minimum: 0
        '403':
          $ref: '#/components/responses/Problem'
      x-infranexum-pagination: offset
      x-infranexum-permission:
        mode: permission
        code: dcim.model.read
    post:
      tags:
      - DCIM / Physical / Models
      operationId: createDcimEquipmentModel
      x-infranexum-capability: dcim.physical
      summary: Create a multi-vendor equipment model and port templates
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEquipmentModel'
      responses:
        '201':
          description: Created model
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentModel'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
      x-infranexum-permission:
        mode: permission
        code: dcim.model.create
  /api/v1/dcim/equipment-taxonomy:
    get:
      tags:
      - DCIM / Physical / Models
      operationId: getDcimEquipmentTaxonomy
      x-infranexum-capability: dcim.physical
      summary: Get governed equipment categories, dependent types and cable types
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Equipment and cable taxonomy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentTaxonomy'
        '403':
          $ref: '#/components/responses/Problem'
      x-infranexum-permission:
        mode: permission
        code: dcim.model.read
  /api/v1/dcim/equipment-models/{id}/status:
    post:
      tags:
      - DCIM / Physical / Models
      operationId: changeDcimEquipmentModelStatus
      x-infranexum-capability: dcim.physical
      summary: Change equipment model lifecycle status
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/Status'
      responses:
        '200':
          description: Updated model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentModel'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-infranexum-permission:
        mode: permission
        code: dcim.model.update
  /api/v1/dcim/racks:
    get:
      tags:
      - DCIM / Physical / Racks
      operationId: listDcimRacks
      x-infranexum-capability: dcim.physical
      summary: List racks, optionally within a room
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: room_id
        in: query
        schema:
          $ref: '#/components/schemas/Uuid'
      - $ref: '#/components/parameters/Limit'
      - name: cursor
        in: query
        required: false
        description: Opaque continuation position encoded as the last returned InfraNexum UUIDv7 identifier.
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Racks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rack'
          headers:
            X-Page-Limit:
              description: Effective bounded page size.
              schema:
                type: integer
                minimum: 1
            X-Next-Cursor:
              description: Cursor for the next page; absent when the current page is terminal.
              schema:
                type: string
                format: uuid
        '403':
          $ref: '#/components/responses/Problem'
      x-infranexum-pagination: cursor
      x-infranexum-permission:
        mode: permission
        code: dcim.rack.read
    post:
      tags:
      - DCIM / Physical / Racks
      operationId: createDcimRack
      x-infranexum-capability: dcim.physical
      summary: Create a rack in an existing room
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRack'
      responses:
        '201':
          description: Created rack
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rack'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
      x-infranexum-permission:
        mode: permission
        code: dcim.rack.create
  /api/v1/dcim/racks/{id}/status:
    post:
      tags:
      - DCIM / Physical / Racks
      operationId: changeDcimRackStatus
      x-infranexum-capability: dcim.physical
      summary: Change rack lifecycle status
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/Status'
      responses:
        '200':
          description: Updated rack
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rack'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-infranexum-permission:
        mode: conditional
        codes:
        - dcim.rack.update
        - dcim.rack.decommission
  /api/v1/dcim/equipment:
    get:
      tags:
      - DCIM / Physical / Equipment
      operationId: listDcimEquipment
      x-infranexum-capability: dcim.physical
      summary: List installed equipment
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: rack_id
        in: query
        schema:
          $ref: '#/components/schemas/Uuid'
      - $ref: '#/components/parameters/Limit'
      - name: cursor
        in: query
        required: false
        description: Opaque continuation position encoded as the last returned InfraNexum UUIDv7 identifier.
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipment'
          headers:
            X-Page-Limit:
              description: Effective bounded page size.
              schema:
                type: integer
                minimum: 1
            X-Next-Cursor:
              description: Cursor for the next page; absent when the current page is terminal.
              schema:
                type: string
                format: uuid
        '403':
          $ref: '#/components/responses/Problem'
      x-infranexum-pagination: cursor
      x-infranexum-permission:
        mode: permission
        code: dcim.equipment.read
    post:
      tags:
      - DCIM / Physical / Equipment
      operationId: installDcimEquipment
      x-infranexum-capability: dcim.physical
      summary: Install equipment into free rack units and instantiate physical ports
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallEquipment'
      responses:
        '201':
          description: Installed equipment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
      x-infranexum-permission:
        mode: permission
        code: dcim.equipment.create
  /api/v1/dcim/equipment/{id}/move:
    post:
      tags:
      - DCIM / Physical / Equipment
      operationId: moveDcimEquipment
      x-infranexum-capability: dcim.physical
      summary: Move equipment to another free rack footprint
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveEquipment'
      responses:
        '200':
          description: Moved equipment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-infranexum-permission:
        mode: permission
        code: dcim.equipment.move
  /api/v1/dcim/equipment/{id}/status:
    post:
      tags:
      - DCIM / Physical / Equipment
      operationId: changeDcimEquipmentStatus
      x-infranexum-capability: dcim.physical
      summary: Change equipment lifecycle status
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/Status'
      responses:
        '200':
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-infranexum-permission:
        mode: conditional
        codes:
        - dcim.equipment.update
        - dcim.equipment.decommission
  /api/v1/dcim/equipment/{id}/ports:
    get:
      tags:
      - DCIM / Physical / Equipment
      operationId: listDcimEquipmentPorts
      x-infranexum-capability: dcim.physical
      summary: List instantiated physical ports for equipment
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Id'
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          maximum: 1000000
      responses:
        '200':
          description: Physical ports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhysicalPort'
          headers:
            X-Page-Limit:
              description: Effective bounded page size.
              schema:
                type: integer
                minimum: 1
            X-Next-Offset:
              description: Offset for the next page; absent when the current page is terminal.
              schema:
                type: integer
                minimum: 0
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
      x-infranexum-pagination: offset
      x-infranexum-permission:
        mode: permission
        code: dcim.port.read
  /api/v1/dcim/cables:
    get:
      tags:
      - DCIM / Physical / Cabling
      operationId: listDcimCables
      x-infranexum-capability: dcim.physical
      summary: List physical cable connections
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - $ref: '#/components/parameters/Limit'
      - name: cursor
        in: query
        required: false
        description: Opaque continuation position encoded as the last returned InfraNexum UUIDv7 identifier.
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Cables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cable'
          headers:
            X-Page-Limit:
              description: Effective bounded page size.
              schema:
                type: integer
                minimum: 1
            X-Next-Cursor:
              description: Cursor for the next page; absent when the current page is terminal.
              schema:
                type: string
                format: uuid
        '403':
          $ref: '#/components/responses/Problem'
      x-infranexum-pagination: cursor
      x-infranexum-permission:
        mode: permission
        code: dcim.cable.read
    post:
      tags:
      - DCIM / Physical / Cabling
      operationId: connectDcimPorts
      x-infranexum-capability: dcim.physical
      summary: Connect two compatible free physical ports
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectPorts'
      responses:
        '201':
          description: Created cable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cable'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
      x-infranexum-permission:
        mode: permission
        code: dcim.cable.create
  /api/v1/dcim/cables/{id}/disconnect:
    post:
      tags:
      - DCIM / Physical / Cabling
      operationId: disconnectDcimCable
      x-infranexum-capability: dcim.physical
      summary: Disconnect an active physical cable
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/Status'
      responses:
        '200':
          description: Disconnected cable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cable'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-infranexum-permission:
        mode: permission
        code: dcim.cable.disconnect
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: INFRANEXUM_SESSION
  parameters:
    OrganizationId:
      name: organization_id
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    Id:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 100
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: ^"ver-[1-9][0-9]*"$
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Caller-generated retry key scoped to the authenticated actor and operation. Reuse with different request semantics
        returns HTTP 409.
      schema:
        type: string
        minLength: 8
        maxLength: 200
        pattern: ^[A-Za-z0-9._:-]+$
  headers:
    ETag:
      schema:
        type: string
    CorrelationId:
      description: Canonical UUIDv7 correlation identifier assigned to the request.
      schema:
        type: string
        format: uuid
        pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
  requestBodies:
    Status:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StatusRequest'
  responses:
    Problem:
      description: RFC 9457 problem
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
  schemas:
    Uuid:
      type: string
      format: uuid
...
    CreateEquipmentModel:
      type: object
      additionalProperties: false
      required:
      - organizationId
      - manufacturerPartnerId
      - displayName
      - formFactor
      - reason
      properties:
        organizationId:
          $ref: '#/components/schemas/Uuid'
        manufacturerPartnerId:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
          minLength: 3
          maxLength: 64
          description: Optional compatibility override. Normal UI creation omits this field and InfraNexum generates a memorable code.
        displayName:
          type: string
          minLength: 3
          maxLength: 128
        category:
          type: string
          description: Equipment category. Use the equipment-taxonomy endpoint to discover supported values.
        equipmentType:
          type: string
          description: Equipment type constrained by category. Use the equipment-taxonomy endpoint to discover supported pairs.
        manufacturerReference:
          type:
          - string
          - 'null'
          maxLength: 160
        formFactor:
          type: string
          maxLength: 32
        rackUnits:
          type: integer
          minimum: 0
          maximum: 100
          description: Required and positive for rack-mountable types; zero is valid for non-rack objects.
        widthMm:
          type: integer
          minimum: 0
          maximum: 5000
        depthMm:
          type: integer
          minimum: 0
          maximum: 5000
        weightKg:
          type:
          - number
          - 'null'
          minimum: 0
        portTemplates:
          type: array
          maxItems: 64
          description: Zero or more port families. Each family may define a different kind/media/connector and count.
          items:
            $ref: '#/components/schemas/PortTemplate'
        description:
          type:
          - string
          - 'null'
          maxLength: 4000
        reason:
          $ref: '#/components/schemas/Reason'
...
    EquipmentTaxonomy:
      type: object
      required: [categories, cableTypes]
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentCategory'
        cableTypes:
          type: array
          items:
            type: string
...
    EquipmentModel:
      type: object
      required:
      - id
      - organizationId
      - manufacturerPartnerId
      - code
      - displayName
      - category
      - equipmentType
      - rackMountable
      - formFactor
      - rackUnits
      - widthMm
      - depthMm
      - portTemplates
      - status
      - version
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        organizationId:
          $ref: '#/components/schemas/Uuid'
        manufacturerPartnerId:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
        displayName:
          type: string
        category:
          type: string
        equipmentType:
          type: string
        rackMountable:
          type: boolean
        manufacturerReference:
          type:
          - string
          - 'null'
        formFactor:
          type: string
        rackUnits:
          type: integer
        widthMm:
          type: integer
        depthMm:
          type: integer
        weightKg:
          type:
          - number
          - 'null'
        portTemplates:
          type: array
          items:
            $ref: '#/components/schemas/PortTemplate'
        status:
          type: string
        description:
          type:
          - string
          - 'null'
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastReason:
          type: string
...